rename networking to stm_networking to force corrrect resolution.
